fix(backend): Return all the admin settings for saveAdminSettings
Solves the null info returned if we only save a part of the admin settings and requested another part of the admin config as result.

For example : save instanceName, and request instanceName and contact in your response. Before : instanceName has the good value, but contact is wrongly null. After : instanceName and contact have the good and updated values.